PestDetectorApp/backend/requirements.txt
torch
torchvision
Pillow
Flask
Flask-Cors
numpy